Move QuestionAnswerer out of test directory, and into non-test directory in services.question package.
