+++ server/app/services/question/QuestionAnswerer.java
@@ -1,4 +1,4 @@
-package support;
+package services.question;
 
 import com.google.common.collect.ImmutableList;
 import services.Path;
