+++ server/test/services/applicant/question/FileUploadQuestionTest.java
@@ -12,9 +12,9 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import services.LocalizedStrings;
 import services.applicant.ApplicantData;
+import services.question.QuestionAnswerer;
 import services.question.types.FileUploadQuestionDefinition;
 import services.question.types.QuestionDefinitionConfig;
-import support.QuestionAnswerer;
 
 @RunWith(JUnitParamsRunner.class)
 public class FileUploadQuestionTest {
