+++ server/test/services/applicant/question/SingleSelectQuestionTest.java
@@ -12,11 +12,11 @@ import org.junit.Test;
 import services.LocalizedStrings;
 import services.applicant.ApplicantData;
 import services.question.LocalizedQuestionOption;
+import services.question.QuestionAnswerer;
 import services.question.QuestionOption;
 import services.question.types.MultiOptionQuestionDefinition;
 import services.question.types.MultiOptionQuestionDefinition.MultiOptionQuestionType;
 import services.question.types.QuestionDefinitionConfig;
-import support.QuestionAnswerer;
 
 public class SingleSelectQuestionTest {
