+++ server/test/services/applicant/question/TextQuestionTest.java
@@ -21,9 +21,9 @@ import services.LocalizedStrings;
 import services.Path;
 import services.applicant.ApplicantData;
 import services.applicant.ValidationErrorMessage;
+import services.question.QuestionAnswerer;
 import services.question.types.QuestionDefinitionConfig;
 import services.question.types.TextQuestionDefinition;
-import support.QuestionAnswerer;
 
 @RunWith(JUnitParamsRunner.class)
 public class TextQuestionTest extends ResetPostgres {
