+++ server/test/services/question/QuestionAnswererTest.java
@@ -1,4 +1,4 @@
-package support;
+package services.question;
 
 import static org.assertj.core.api.Assertions.assertThat;
