+++ server/test/views/questiontypes/AddressRendererTest.java
@@ -16,9 +16,9 @@ import repository.ResetPostgres;
 import services.LocalizedStrings;
 import services.applicant.ApplicantData;
 import services.applicant.question.ApplicantQuestion;
+import services.question.QuestionAnswerer;
 import services.question.types.AddressQuestionDefinition;
 import services.question.types.QuestionDefinitionConfig;
-import support.QuestionAnswerer;
 
 public class AddressRendererTest extends ResetPostgres {
